+++ src/game/scripts/vscripts/ingame.lua
@@ -326,16 +326,18 @@ dc_table = {};
 function Ingame:onStart()
     local this = self
 
-    -- Force bots to take a defensive pose for the first 5 minutes of the game. This is top stop bots from straight away pushing lanes when they hit level 6
+    -- Force bots to take a defensive pose until the first tower has been destroyed. This is top stop bots from straight away pushing lanes when they hit level 6
     Timers:CreateTimer(function ()
                Convars:SetBool("dota_tutorial_force_bot_defend", true)
                print("bots will only defend")
             end, 'forceBotsToDefend', 2)
 
-    Timers:CreateTimer(function ()
-               Convars:SetBool("dota_tutorial_force_bot_defend", false)
-               print("bots will now push")
-            end, 'letBotsAttack', 480)
+
+   -- This is an alternative restart condition that reverts bots after 8 minutes, however, a lot damage can be done in 8 minutes.
+   -- Timers:CreateTimer(function ()
+   --            Convars:SetBool("dota_tutorial_force_bot_defend", false)
+   --            print("bots will now push")
+   --         end, 'letBotsAttack', 480)
     
     ---Enable and then quickly disable all vision. This fixes two problems. First it fixes the scoreboard missing enemy abilities, and second it fixes the issues of bots not moving until they see an enemy player.
     if Convars:GetBool("dota_all_vision") == false then
@@ -1363,6 +1365,9 @@ function Ingame:addStrongTowers()
         end
     end, nil)
     ListenToGameEvent('dota_tower_kill', function (keys)
+        -- If a tower is destroyed, return bots to their normal behaviour
+        Convars:SetBool("dota_tutorial_force_bot_defend", false)
+        
         if OptionManager:GetOption('strongTowers') then
             local tower_team = keys.teamnumber
             local towers = Entities:FindAllByClassname('npc_dota_tower')